Added libc headers and some string.h impl

+++ libc/string/strlen.c
@@ -1,6 +1,6 @@
 #include <string.h>
 
-int strlen(const char *str) {
+size_t strlen(const char *str) {
     const char *s;
     for(s = str; *s; ++s);
     return s - str;
